+++ src/main/java/survivalblock/laseredstone/common/package-info.java
@@ -1,0 +1,8 @@
+@MethodsReturnNonnullByDefault
+@FieldsAreNonnullByDefault
+@NullMarked
+package survivalblock.laseredstone.common;
+
+import com.mojang.logging.annotations.FieldsAreNonnullByDefault;
+import com.mojang.logging.annotations.MethodsReturnNonnullByDefault;
+import org.jspecify.annotations.NullMarked;